update world tree
