+++ bin/tree_availability_service.rs
@@ -16,6 +16,7 @@ use world_tree::tree::config::{ServiceConfig, WorldTreeConfig};
 use world_tree::tree::identity_tree::{self, IdentityTree};
 use world_tree::tree::service::InclusionProofService;
 use world_tree::tree::tree_manager::{BridgedTree, CanonicalTree, TreeManager};
+use world_tree::tree::WorldTree;
 
 /// This service syncs the state of the World Tree and spawns a server that can deliver inclusion proofs for a given identity.
 #[derive(Parser, Debug)]
@@ -50,9 +51,9 @@ pub async fn main() -> eyre::Result<()> {
         init_subscriber(Level::INFO);
     }
 
-    let identity_tree = initialize_identity_tree(&config.world_tree).await?;
+    let world_tree = initialize_identity_tree(&config.world_tree).await?;
 
-    let handles = InclusionProofService::new(identity_tree)
+    let handles = InclusionProofService::new(world_tree)
         .serve(config.world_tree.socket_address);
 
     let mut handles = handles.into_iter().collect::<FuturesUnordered<_>>();
@@ -68,7 +69,7 @@ pub async fn main() -> eyre::Result<()> {
 
 async fn initialize_identity_tree(
     config: &WorldTreeConfig,
-) -> eyre::Result<IdentityTree<Provider<Http>>> {
+) -> eyre::Result<WorldTree<Provider<Http>>> {
     let http_provider =
         Http::new(config.canonical_tree.provider.rpc_endpoint.clone());
     let canonical_middleware = Arc::new(Provider::new(http_provider));
@@ -99,7 +100,7 @@ async fn initialize_identity_tree(
         bridged_tree_managers.push(tree_manager);
     }
 
-    Ok(IdentityTree::new(
+    Ok(WorldTree::new(
         config.tree_depth,
         canonical_tree_manager,
         bridged_tree_managers,
